+++ Magazzino/src/Prodotti/Tablet.java
@@ -1,5 +1,7 @@
 package Prodotti;
 
+import java.math.BigDecimal;
+
 public class Tablet extends Prodotto {
 
     public Tablet(){
@@ -7,24 +9,9 @@ public class Tablet extends Prodotto {
         this.tipo = TipoProdotto.TABLET;
     }
 
-    public Tablet(TipoProdotto tipo, String produttore, String modello, String id, String descrizione, double dimDisplay, int dimMemoria, double prezzoAcquisto, double prezzoVendita) {
+    public Tablet(TipoProdotto tipo, String produttore, String modello, String id, String descrizione,
+                  double dimDisplay, int dimMemoria, BigDecimal prezzoAcquisto, BigDecimal prezzoVendita) {
         super(tipo, produttore, modello, id, descrizione, dimDisplay, dimMemoria, prezzoAcquisto, prezzoVendita);
         this.tipo = TipoProdotto.TABLET;
     }
-
-//    @Override
-//    public String toString() {
-//        return "Prodotto{" +
-//                "tipo=" + getTipo() +
-//                ", produttore='" + getProduttore() + '\'' +
-//                ", modello='" + getModello() + '\'' +
-//                ", descrizione='" + getDescrizione() + '\'' +
-//                ", dimDisplay=" + getDimDisplay() +
-//                ", dimMemoria=" + getDimMemoria() +
-//                ", prezzoAcquisto=" + getPrezzoAcquisto() +
-//                ", prezzoVendita=" + getPrezzoVendita() +
-//                ", id='" + getId() + '\'' +
-//                '}' + "\n ";
-//    }
-
 }
